+++ src/JSystem/JKernel/JKRAram.cpp
@@ -99,7 +99,7 @@ void* JKRAram::run()
 void JKRAram::checkOkAddress(u8* addr, u32 size, JKRAramBlock* block,
                              u32 param_4)
 {
-	if (!IS_ALIGNED((u32)addr, 0x20) && !IS_ALIGNED(size, 0x20)) {
+	if (!IS_ALIGNED((uintptr_t)addr, 0x20) && !IS_ALIGNED(size, 0x20)) {
 		OSPanic(__FILE__, 225, ":::address not 32Byte aligned.");
 	}
 
@@ -157,7 +157,7 @@ JKRAramBlock* JKRAram::mainRamToAram(u8* buf, u32 bufSize, u32 alignedSize,
 			block = nullptr;
 		} else {
 			JKRDecompress(buf, (u8*)allocatedMem, fileSize, 0);
-			JKRAramPcs(0, (u32)allocatedMem, bufSize, alignedSize, block);
+			JKRAramPcs(0, (u32)(uintptr_t)allocatedMem, bufSize, alignedSize, block);
 			JKRFreeToHeap(heap, allocatedMem);
 			block = block == nullptr ? (JKRAramBlock*)-1 : block;
 		}
@@ -174,7 +174,7 @@ JKRAramBlock* JKRAram::mainRamToAram(u8* buf, u32 bufSize, u32 alignedSize,
 			bufSize = allocatedBlock->getAddress();
 		}
 
-		JKRAramPcs(0, (u32)buf, bufSize, alignedSize, block);
+		JKRAramPcs(0, (u32)(uintptr_t)buf, bufSize, alignedSize, block);
 		block = block == nullptr ? (JKRAramBlock*)-1 : block;
 	}
 	return block;
@@ -212,8 +212,8 @@ u8* JKRAram::aramToMainRam(u32 address, u8* buf, u32 p3,
 	u32 expandSize;
 	if (expandSwitch == EXPAND_SWITCH_DECOMPRESS) {
 		u8 buffer[64];
-		u8* bufPtr = (u8*)ALIGN_NEXT((u32)buffer, 32);
-		JKRAramPcs(1, address, (u32)bufPtr, 32, nullptr);
+		u8* bufPtr = (u8*)ALIGN_NEXT((uintptr_t)buffer, 32);
+		JKRAramPcs(1, address, (u32)(uintptr_t)bufPtr, 32, nullptr);
 		compression = JKRCheckCompressed(bufPtr);
 		expandSize  = JKRDecompExpandSize(bufPtr);
 	}
@@ -239,7 +239,7 @@ u8* JKRAram::aramToMainRam(u32 address, u8* buf, u32 p3,
 		if (szpSpace == nullptr)
 			return nullptr;
 
-		JKRAramPcs(1, address, (u32)szpSpace, p3, nullptr);
+		JKRAramPcs(1, address, (u32)(uintptr_t)szpSpace, p3, nullptr);
 		if (p5 != 0 && p5 < expandSize)
 			expandSize = p5;
 
@@ -265,7 +265,7 @@ u8* JKRAram::aramToMainRam(u32 address, u8* buf, u32 p3,
 			return nullptr;
 
 		changeGroupIdIfNeed(rv, id);
-		JKRAramPcs(1, address, (u32)rv, p3, nullptr);
+		JKRAramPcs(1, address, (u32)(uintptr_t)rv, p3, nullptr);
 		if (pSize != nullptr) {
 			*pSize = p3;
 		}
@@ -462,7 +462,7 @@ static u8* firstSrcData()
 	u32 size   = szpEnd - buffer;
 	u32 length = transLeft < size ? transLeft : size;
 
-	JKRAramPcs(1, srcAddress + srcOffset, (u32)buffer, ALIGN_NEXT(length, 0x20),
+	JKRAramPcs(1, srcAddress + srcOffset, (u32)(uintptr_t)buffer, ALIGN_NEXT(length, 0x20),
 	           nullptr);
 
 	srcOffset += length;
@@ -488,7 +488,7 @@ static u8* nextSrcData(u8* current)
 	}
 	JUT_ASSERT(VERSION_SELECT(1376, 1361, 1361), transSize > 0);
 
-	JKRAramPcs(1, (u32)(srcAddress + srcOffset), ((u32)dest + left),
+	JKRAramPcs(1, (u32)(srcAddress + srcOffset), ((u32)(uintptr_t)dest + left),
 	           ALIGN_NEXT(transSize, 0x20), nullptr);
 	srcOffset += transSize;
 	transLeft -= transSize;
