+++ src/JSystem/JKernel/JKRAramStream.cpp
@@ -104,7 +104,7 @@ s32 JKRAramStream::writeToAram(JKRAramStreamCommand* command)
 
 			s32 readLength = command->mStream->read(buffer, length);
 
-			JKRAramPcs(0, (u32)buffer, destination, length, nullptr);
+			JKRAramPcs(0, (u32)(uintptr_t)buffer, destination, length, nullptr);
 			dstSize -= length;
 			offset += length;
 			writtenLength += length;
@@ -183,7 +183,7 @@ void JKRAramStream::setTransBuffer(u8* buffer, u32 bufferSize, JKRHeap* heap)
 	transHeap   = nullptr;
 
 	if (buffer) {
-		transBuffer = (u8*)ALIGN_NEXT((u32)buffer, 0x20);
+		transBuffer = (u8*)ALIGN_NEXT((uintptr_t)buffer, 0x20);
 	}
 
 	if (bufferSize) {
